requirements-docker.txt
# LangChain
langchain>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.0
langchain-huggingface>=0.1.0
langgraph>=0.2.0
# HuggingFace (torch/torchvision installed separately in Dockerfile for CPU wheel URL)
huggingface-hub>=0.23.0
transformers>=4.40.0
# Streamlit + UI
streamlit>=1.35.0
streamlit-folium>=0.20.0
# Mapping / charts
folium>=0.16.0
plotly>=5.20.0
# Data / HTTP
pandas>=2.2.0
requests>=2.31.0
# Config / validation
python-dotenv>=1.0.0
pydantic>=2.7.0
